Rename pinky folder in core folder
