+++ src/core/CMakeLists.txt
