+++ src/core/constant.hpp
