+++ src/core/error.hpp
