+++ src/core/function.cpp
