+++ src/core/operator.cpp
