+++ src/core/processing.cpp
