+++ tests/test_tokenization.cpp
@@ -1,6 +1,6 @@
 #include "catch2/catch.hpp"
 
-#include "pinky/processing.hpp"
+#include "core/processing.hpp"
 
 
 TEST_CASE("tokenization - expression with one number", "[processing]")
